requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
protobuf==3.18.1
python-dotenv==0.19.1
typing==3.7.4.3
Werkzeug==2.0.1
